requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0

# SQLAlchemy async + Supabase Postgres
SQLAlchemy==2.0.29
asyncpg==0.29.0

# Autenticación
python-jose==3.3.0
cryptography==42.0.5

# Configuración y env
pydantic==2.7.1
pydantic-settings==2.2.1

# Utils
python-multipart==0.0.9

# Para logging más pro (opcional)
loguru==0.7.2
